Don't filter `listWorkflowRuns` using `head_sha`.
This commit removes the filtering of `listWorkflowRuns` by the
`head_sha` param as it appears to be broken upstream. See
https://github.com/dawidd6/action-download-artifact/issues/243.

Instead pull a list of workflow runs and then filter them based on the
`head_sha` value.

Also tweaked behaviour to resolve the `workflow` to an ID, and added
some additional debug logging.